+++ magda-minion-linked-data-rating/src/datasetQualityAspectDef.ts
