+++ magda-minion-linked-data-rating/src/onRecordFound.ts
@@ -4,7 +4,7 @@ import { Record } from "magda-typescript-common/src/generated/registry/api";
 import unionToThrowable from "magda-typescript-common/src/util/unionToThrowable";
 import AuthorizedRegistryClient from "magda-typescript-common/src/registry/AuthorizedRegistryClient";
 import linkedDataAspectDef from "./linkedDataAspectDef";
-import datasetQualityAspectDef from "magda-minion-framework/src/common-aspect-defs/datasetQualityAspectDef";
+import datasetQualityAspectDef from "./datasetQualityAspectDef";
 import openLicenses from "./openLicenses";
 import formatStars from "./openFormats";
