+++ magda-typescript-common/README.md
@@ -1,0 +1,17 @@
+### Magda Typescript Common
+
+A Common-JS module contains common magda utils.
+
+#### Installation
+
+```bash
+yarn add @magda/typescript-common
+```
+
+#### Usage Examples
+
+```javascript
+import RegistryClient from "@magda/typescript-common/dist/registry/RegistryClient";
+```
+
+To find out what utils are available through this package, please check [here](https://github.com/magda-io/magda/tree/master/magda-typescript-common/src)
